test: addressed review comment
