+++ db/indextest/indextest_test.go
@@ -12,10 +12,13 @@ package indextest
 
 import (
 	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/couchbase/sync_gateway/base"
+	"github.com/couchbase/sync_gateway/channels"
 	"github.com/couchbase/sync_gateway/db"
+	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -39,18 +42,18 @@ func TestRoleQuery(t *testing.T) {
 		t.Run(fmt.Sprintf("TestRoleQuery in Serverless=%t", testCase.isServerless), func(t *testing.T) {
 			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
 
-			db, ctx := setupTestDBForBucketWithOptions(t, dbContextConfig)
-			defer db.Close(ctx)
+			database, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
+			defer database.Close(ctx)
 
-			n1QLStore, reset, err := setupN1QLStore(db.Bucket, testCase.isServerless)
+			n1QLStore, reset, err := setupN1QLStore(database.Bucket, testCase.isServerless)
 			assert.NoError(t, err, "Unable to get n1QLStore for testBucket")
 			defer func(n1QLStore base.N1QLStore, isServerless bool) {
 				err := reset(n1QLStore, isServerless)
 				assert.NoError(t, err, "Reset fn shouldn't return error")
 			}(n1QLStore, testCase.isServerless)
 
-			authenticator := db.Authenticator(ctx)
-			assert.NotNil(t, authenticator, "db.Authenticator(ctx) returned nil")
+			authenticator := database.Authenticator(ctx)
+			assert.NotNil(t, authenticator, "database.Authenticator(ctx) returned nil")
 
 			// Add roles
 			for i := 1; i <= 5; i++ {
@@ -66,7 +69,7 @@ func TestRoleQuery(t *testing.T) {
 			assert.NoError(t, err)
 
 			// Standard query
-			results, queryErr := db.QueryRoles(ctx, "", 0)
+			results, queryErr := database.QueryRoles(ctx, "", 0)
 			assert.NoError(t, queryErr, "Query error")
 			var row map[string]interface{}
 			rowCount := 0
@@ -100,7 +103,7 @@ func TestBuildRolesQuery(t *testing.T) {
 		t.Run(fmt.Sprintf("BuildRolesQuery in Serverless=%t", testCase.isServerless), func(t *testing.T) {
 			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
 
-			database, ctx := setupTestDBForBucketWithOptions(t, dbContextConfig)
+			database, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
 			defer database.Close(ctx)
 
 			n1QLStore, reset, err := setupN1QLStore(database.Bucket, testCase.isServerless)
@@ -143,7 +146,7 @@ func TestBuildSessionsQuery(t *testing.T) {
 		t.Run(fmt.Sprintf("BuildSessionsQuery in Serverless=%t", testCase.isServerless), func(t *testing.T) {
 			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
 
-			database, ctx := setupTestDBForBucketWithOptions(t, dbContextConfig)
+			database, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
 			defer database.Close(ctx)
 
 			n1QLStore, reset, err := setupN1QLStore(database.Bucket, testCase.isServerless)
@@ -186,7 +189,7 @@ func TestBuildUsersQuery(t *testing.T) {
 		t.Run(fmt.Sprintf("TestBuildUsersQuery in Serverless=%t", testCase.isServerless), func(t *testing.T) {
 			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
 
-			database, ctx := setupTestDBForBucketWithOptions(t, dbContextConfig)
+			database, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
 			defer database.Close(ctx)
 
 			n1QLStore, reset, err := setupN1QLStore(database.Bucket, testCase.isServerless)
@@ -229,7 +232,7 @@ func TestQueryAllRoles(t *testing.T) {
 		t.Run(fmt.Sprintf("TestQueryAllRoles in Serverless=%t", testCase.isServerless), func(t *testing.T) {
 			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
 
-			database, ctx := setupTestDBForBucketWithOptions(t, dbContextConfig)
+			database, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
 			defer database.Close(ctx)
 
 			n1QLStore, reset, err := setupN1QLStore(database.Bucket, testCase.isServerless)
@@ -268,4 +271,193 @@ func TestQueryAllRoles(t *testing.T) {
 			assert.NoError(t, results.Close())
 		})
 	}
+}
+
+func TestAllPrincipalIDs(t *testing.T) {
+	if base.UnitTestUrlIsWalrus() || base.TestsDisableGSI() {
+		t.Skip("This test is Couchbase Server and UseViews=false only")
+	}
+
+	testCases := []struct {
+		isServerless bool
+	}{
+		{
+			isServerless: false,
+		},
+		{
+			isServerless: true,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(fmt.Sprintf("TestAllPrincipalIDs in Serverless=%t", testCase.isServerless), func(t *testing.T) {
+			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
+
+			db, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
+			defer db.Close(ctx)
+
+			n1QLStore, reset, err := setupN1QLStore(db.Bucket, testCase.isServerless)
+			assert.NoError(t, err, "Unable to get n1QLStore for testBucket")
+			defer func(n1QLStore base.N1QLStore, isServerless bool) {
+				err := reset(n1QLStore, isServerless)
+				assert.NoError(t, err, "Reset fn shouldn't return error")
+			}(n1QLStore, testCase.isServerless)
+			base.SetUpTestLogging(t, base.LevelDebug, base.KeyCache, base.KeyChanges)
+
+			db.Options.QueryPaginationLimit = 100
+			db.ChannelMapper = channels.NewDefaultChannelMapper()
+			authenticator := db.Authenticator(ctx)
+
+			rolename1 := uuid.NewString()
+			rolename2 := uuid.NewString()
+			username := uuid.NewString()
+
+			user1, err := authenticator.NewUser(username, "letmein", nil)
+			assert.NoError(t, err)
+			user1.SetExplicitRoles(channels.TimedSet{rolename1: channels.NewVbSimpleSequence(1), rolename2: channels.NewVbSimpleSequence(1)}, 1)
+			assert.NoError(t, authenticator.Save(user1))
+
+			role1, err := authenticator.NewRole(rolename1, nil)
+			assert.NoError(t, err)
+			assert.NoError(t, authenticator.Save(role1))
+
+			role2, err := authenticator.NewRole(rolename2, nil)
+			assert.NoError(t, err)
+			assert.NoError(t, authenticator.Save(role2))
+
+			err = db.DeleteRole(ctx, role2.Name(), false)
+			assert.NoError(t, err)
+			roleGet, err := authenticator.GetRoleIncDeleted(role2.Name())
+			assert.NoError(t, err)
+			assert.True(t, roleGet.IsDeleted())
+
+			t.Log("user1:", user1.Name())
+			t.Log("role1:", role1.Name())
+			t.Log("role2:", role2.Name())
+
+			// assert allprincipals still returns users and deleted roles
+			users, roles, err := db.AllPrincipalIDs(ctx)
+			assert.NoError(t, err)
+			assert.ElementsMatch(t, []string{user1.Name()}, users)
+			assert.ElementsMatch(t, []string{role1.Name(), role2.Name()}, roles)
+
+			roles, err = db.GetRoleIDs(ctx, false)
+			assert.NoError(t, err)
+			assert.ElementsMatch(t, []string{role1.Name()}, roles)
+		})
+	}
+}
+
+func TestGetRoleIDs(t *testing.T) {
+	if base.UnitTestUrlIsWalrus() || base.TestsDisableGSI() {
+		t.Skip("This test is Couchbase Server and UseViews=false only")
+	}
+
+	testCases := []struct {
+		isServerless   bool
+		includeDeleted bool
+	}{
+		{
+			isServerless:   false,
+			includeDeleted: false,
+		},
+		{
+			isServerless:   false,
+			includeDeleted: true,
+		},
+		{
+			isServerless:   true,
+			includeDeleted: false,
+		},
+		{
+			isServerless:   true,
+			includeDeleted: true,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(fmt.Sprintf("TestGetRoleIDs {Serverless=%t; includeDelete=%t}", testCase.isServerless, testCase.includeDeleted), func(t *testing.T) {
+			dbContextConfig := getDatabaseContextOptions(testCase.isServerless)
+
+			database, ctx := db.SetupTestDBWithOptions(t, dbContextConfig)
+			defer database.Close(ctx)
+
+			n1QLStore, reset, err := setupN1QLStore(database.Bucket, testCase.isServerless)
+			assert.NoError(t, err, "Unable to get n1QLStore for testBucket")
+			defer func(n1QLStore base.N1QLStore, isServerless bool) {
+				err := reset(n1QLStore, isServerless)
+				assert.NoError(t, err, "Reset fn shouldn't return error")
+			}(n1QLStore, testCase.isServerless)
+			base.SetUpTestLogging(t, base.LevelDebug, base.KeyCache, base.KeyChanges)
+
+			database.Options.QueryPaginationLimit = 100
+			database.ChannelMapper = channels.NewDefaultChannelMapper()
+			authenticator := database.Authenticator(ctx)
+
+			rolename1 := uuid.NewString()
+			rolename2 := uuid.NewString()
+
+			role1, err := authenticator.NewRole(rolename1, nil)
+			assert.NoError(t, err)
+			assert.NoError(t, authenticator.Save(role1))
+
+			role2, err := authenticator.NewRole(rolename2, nil)
+			assert.NoError(t, err)
+			assert.NoError(t, authenticator.Save(role2))
+
+			err = database.DeleteRole(ctx, role2.Name(), false)
+			assert.NoError(t, err)
+			roleGet, err := authenticator.GetRoleIncDeleted(role2.Name())
+			assert.NoError(t, err)
+			assert.True(t, roleGet.IsDeleted())
+
+			t.Log("role1:", role1.Name())
+			t.Log("role2:", role2.Name())
+
+			// assert roles
+			roles, err := database.GetRoleIDs(ctx, testCase.includeDeleted)
+			expectedRoles := []string{role1.Name()}
+			if testCase.includeDeleted {
+				expectedRoles = append(expectedRoles, role2.Name())
+			}
+
+			assert.NoError(t, err)
+			assert.ElementsMatch(t, expectedRoles, roles)
+		})
+	}
+}
+
+func getDatabaseContextOptions(isServerless bool) db.DatabaseContextOptions {
+	defaultCacheOptions := db.DefaultCacheOptions()
+
+	return db.DatabaseContextOptions{
+		CacheOptions: &defaultCacheOptions,
+		Serverless:   isServerless,
+	}
+}
+
+type resetN1QLStoreFn func(n1QLStore base.N1QLStore, isServerless bool) error
+
+func setupN1QLStore(bucket base.Bucket, isServerless bool) (base.N1QLStore, resetN1QLStoreFn, error) {
+	n1QLStore, ok := base.AsN1QLStore(bucket)
+	if !ok {
+		return nil, nil, fmt.Errorf("Unable to get n1QLStore for testBucket")
+	}
+
+	if err := db.InitializeIndexes(n1QLStore, base.TestUseXattrs(), 0, false, isServerless); err != nil {
+		return nil, nil, err
+	}
+
+	return n1QLStore, clearIndexes, nil
+}
+
+var clearIndexes resetN1QLStoreFn = func(n1QLStore base.N1QLStore, isServerless bool) error {
+	indexes := db.GetIndexesName(isServerless, base.TestUseXattrs())
+	for _, index := range indexes {
+		err := n1QLStore.DropIndex(index)
+		if err != nil && strings.Contains(err.Error(), "Index not exist") {
+			return err
+		}
+	}
+	return nil
 }
